created : 위시리스트 구현

+++ sale/src/main/java/kr/game/sale/repository/game/GameRepository.java
@@ -7,5 +7,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import java.util.Optional;
 
 public interface GameRepository extends JpaRepository<Game, Long>, GameRepositoryCustom {
-    Game findBySteamAppid(int steamAppid);
+    Game findBySteamAppid(Long steamAppid);
 }
